[Rename] 컴포넌트 폴더로 묶음
