+++ deco/src/components/LoginPage/Login.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import WelcomeInfo from "@/components/Common/WelcomeInfo";
-import FormInput from "@/components/Common/FormInput";
+import WelcomeInfo from "@/components/Common/WelcomeInfo/WelcomeInfo";
+import FormInput from "@/components/Common/FormInput/FormInput";
 import classes from "./Login.module.css";
 
 const Login = () => {
